refactor(#352): split ClusterLeader.cpp into its TU family; dedupe digest build
ClusterLeader.cpp (1876) → core (statics + init + HTTP helpers + tick) +
Grid (submit/self-row/mirror) + Fanout (config swap, member work,
result fold) + Rollout (#276/#344 dispatch) + FollowerPush (#304 relay +
file-source stream) + Maintenance (#321 reboot-hold, #318E log pull) +
Status (statusFillLocked + staging API), shared seams in
ClusterLeaderInternal.h (family-only include). All files under the
800-line cap.

The duplicated ~17-line digest-freshness sequence (collectMemberWork +
rebootHoldBuildTargets) is now one buildFreshDigestLocked() (#355 item).

CLAUDE.md file pointers updated for both #352 splits. No logic change;
781 native + 77 pytest green.
